updated repository, model, service and rest for customer

+++ src/main/java/com/parqueApp/parqueApp/model/Customer.java
@@ -21,6 +21,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 /**
  * The class represents the customer lot entity.
@@ -39,10 +40,10 @@ public class Customer implements Serializable {
     int id_card;
     String first_name;
     String last_name;
-    int phone;
+    BigInteger phone;
     String address;
 
-    public Customer(int id_card, String first_name, String last_name, int phone, String address, int id_card1, String first_name1, String last_name1, int phone1, String address1) {
+    public Customer(int id_card, String first_name, String last_name, int phone, String address, int id_card1, String first_name1, String last_name1, BigInteger phone1, String address1) {
         this.id_card = id_card1;
         this.first_name = first_name1;
         this.last_name = last_name1;
